requirements.txt
Django==1.11.15
django-environ==0.4.5
psycopg2
djangorestframework
djangorestframework-gis
coreapi
markdown